+++ views/diagnostics-page.php
@@ -25,18 +25,6 @@ if ( ! defined( 'ABSPATH' ) ) {
         </thead>
         <tbody>
             <tr>
-                <td>PHP max_execution_time</td>
-                <td>
-                    <?php echo $view['maxExecutionTime'] === 0 ? 'Unlimited' : esc_html( $view['maxExecutionTime'] ) . ' secs'; ?>
-
-                    <span
-                        class="dashicons <?php echo $view['maxExecutionTime'] === 0 ? 'dashicons-yes' : 'dashicons-no'; ?>"
-                        style="color: <?php echo $view['maxExecutionTime'] === 0 ? 'green' : 'red'; ?>;"
-                    ></span>
-                </td>
-                <td>Generating a static site can involve long-running processes. Set your PHP max_execution_time setting to unlimited or find a better webhost if you're prevented from doing so.</td>
-            </tr>
-            <tr>
                 <td>PHP memory_limit</td>
                 <td>
                     <?php echo esc_html( $view['memoryLimit'] ); ?>